Implements data systems (#82816)
Subsystems currently come in two different flavors:
1. Systems that process at intervals with the master controller
2. Global data containers that do not fire

And I think they should be split up...

This moves 4 non firing, non init subsytems -> datasystem

Clarity in code

+++ code/game/objects/effects/effect_system/effects_explosion.dm
@@ -9,11 +9,15 @@
 	return INITIALIZE_HINT_LATELOAD
 
 /obj/effect/particle_effect/expl_particles/LateInitialize()
-	var/direct = pick(GLOB.alldirs)
-	var/steps_amt = pick(25;1,50;2,100;3,200;4)
-	for(var/j in 1 to steps_amt)
-		step(src, direct)
-		sleep(1)
+	var/step_amt = pick(25;1,50;2,100;3,200;4)
+
+	var/datum/move_loop/loop = DSmove_manager.move(src, pick(GLOB.alldirs), 1, timeout = step_amt, priority = MOVEMENT_ABOVE_SPACE_PRIORITY)
+	RegisterSignal(loop, COMSIG_QDELETING, PROC_REF(end_particle))
+
+/obj/effect/particle_effect/expl_particles/proc/end_particle(datum/source)
+	SIGNAL_HANDLER
+	if(QDELETED(src))
+		return
 	qdel(src)
 
 /datum/effect_system/expl_particles
